% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datasets.R
\name{download_datasets}
\alias{download_datasets}
\title{Download datasets}
\usage{
download_datasets(url, collection, target, method = "auto", replace = FALSE)
}
\arguments{
\item{url}{URL of dataset repository}

\item{collection}{filename of collection under the dataset repository}

\item{target}{target directory for downloading}

\item{method}{a method argument passed to download.file}

\item{replace}{a boolean flag indicating whether to replace existing datasets}
}
\description{
\code{download_datasets} Download datasets
}
